Feat: Finishing update and adding reads and delete routes. allowing log server

+++ repositories/User/UserRepository.py
@@ -7,33 +7,43 @@ class UserRepository:
     def __init__(self):
         self.__entity = User
         self.__model_creator = pydantic_model_creator(User)
-        self.__tortoise = Tortoise.get_connection('default')
+        #self.__tortoise = Tortoise.get_connection('default')
 
-    async def to_dict(self, user: User) -> dict:
+    async def toDict(self, user: User) -> dict:
         result = await self.__model_creator.from_tortoise_orm(user)
         return result.model_dump()
 
     async def create(self, user: dict) -> UserDTO:
         newUser = await self.__entity.create(**user) 
-        return UserDTO.model_validate(await self.to_dict(newUser))
+        return UserDTO.model_validate(await self.toDict(newUser))
     
-    async def update(self, user: User) -> UserDTO:
-        await user.save()
-        return UserDTO.model_validate(await self.to_dict(user))
+    async def update(self, userDTO: UserDTO) -> UserDTO:
+        user = User(**userDTO.model_dump())
+        await user.save(force_update=True)
+        return UserDTO.model_validate(await self.toDict(user))
 
-    async def get_one(self, id: int) -> UserDTO:
+    async def getOne(self, id: int) -> UserDTO:
         user = await self.__entity.get_or_none(id=id)
-        return UserDTO.model_validate(await self.to_dict(user))
+        return UserDTO.model_validate(await self.toDict(user))
     
-    async def get_all(self) -> list[User]:
-        return await self.__entity.all()
+    async def getAll(self) -> list[UserDTO]:
+        users = []
+        for user in await self.__entity.all():
+            users.append(UserDTO.model_validate(await self.toDict(user)))
+        return users
     
-    async def get_many(self, key: str, value: any) -> list[User]:
-        return await self.__entity.filter(**{key:value})
+    async def getMany(self, key: str, value: any) -> list[UserDTO]:
+        users = list[UserDTO]
+        for user in await self.__entity.filter(**{key:value}):
+            users.append(UserDTO.model_validate(await self.toDict(user)))
+        return users
 
-    async def delete(self, user: User) -> bool:
-        await user.delete()
-        return True
+    async def delete(self, id: int) -> bool:
+        user = await self.__entity.get_or_none(id=id)
+        if user:
+            await user.delete()
+            return True
+        return False
     
-    async def execute_sql(self, query: str) -> list[dict]:
-        return await self.__tortoise.execute_query_dict(query)
+    # async def execute_sql(self, query: str) -> list[dict]:
+    #     return await self.__tortoise.execute_query_dict(query)
